+++ package.json
@@ -54,8 +54,7 @@
     "style-loader": "0.13.0",
     "url-loader": "0.5.7",
     "webpack": "1.12.9",
-    "webpack-dev-middleware": "1.4.0",
-    "webpack-hot-middleware": "2.6.0"
+    "webpack-dev-server": "^1.14.1"
   },
   "dependencies": {
     "react": "0.14.6",
